DWEELING
code with name, not nullable and length mx = 10; urlIMage with name and not nullable; area with name, not nillable and length mx = 10; price with name and not nullable; description with name, not nullable and Lob; numRomms with name, not nullable and length mx = 10; numBathrooms with name, not nullable and length mx = 10; type with name and not nullable, this is enum; project with name.
Constructor method from Dweling, document method.